runtime: split PipelineExecutionService into focused collaborators
